make macros with \vv robust
this prevents strange errors when using them within todo commands, such
as \bfside

the problem appears to be that \vv is a fragile command, and todo
commands have moving arguments, which means that the argument of a todo
command is saved to an auxillary file during compilation.
Fragile means that \vv expands to invalid TeX code in this save process,
leading to errors. Making the command robust fixes the issue, then the
command is not expanded before the saving happens and only expanded later.